more refactoring moved cpu gpu and wifi to respective hardware probe modules.
